feat: add sorting by region
