+++ src/App.tsx
@@ -1,4 +1,4 @@
-import Homepage from './components/Homepage'
+import Homepage from './components/Homepage/Homepage'
 
 function App() {
   return (
