+++ src/main.tsx
@@ -1,6 +1,6 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import './index.css'
+import './index.scss'
 import App from './App.tsx'
 
 createRoot(document.getElementById('root')!).render(
